Debug.WriteLine in place of Console.WriteLine
Hello,

I'm using this library in a project which is communicating via CGI with a nodejs project. The latter takes the console output of the C# project in a json format. My problem is that when the iRacing yaml fail to parse via the YamlParser, the error is sent to the standard output via the `Console.WriteLine` which pollute the output and any client connected to the app fails to parse as it's not JSON.

As it is an error message and a stack trace, I propose to write them in the Debug trace listeners instead, that way we can still see them in a debug env.